Fix runtime error in handling error messages response
It seems the parser represent single node as a hash, and multiple nodes
with the same name as an array. According to asendia documentation
StatusDetails is a list / collection of StatusDetail nodes, but the
resulting dictionary in current parser is a bit awkward,
i.e. StatusDetail becomes a list of hashes.

For Example:

<StatusDetails><StatusDetail><Message/></StatusDetail></StatusDetails>
 => ['StatusDetails']['StatusDetail']['Message']

<StatusDetails><StatusDetail><Message/></StatusDetail><StatusDetail><Message/></StatusDetail></StatusDetails>
 => ['StatusDetails']['StatusDetail'][0]['Message']